requirements.txt
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bce-python-sdk==0.9.60
certifi==2026.1.4
chardet==7.0.1
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.0
fastar==0.8.0
filelock==3.25.0
fsspec==2026.2.0
future==1.0.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.5.0
idna==3.11
imagesize==2.0.0
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
modelscope==1.34.0
networkx==3.6.1
numpy==2.4.1
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging==26.0
paddleocr==3.4.0
paddlepaddle==3.2.0
paddlex==3.4.2
pandas==3.0.1
pillow==12.1.0
prettytable==3.17.0
protobuf==6.33.4
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pygments==2.19.2
pymupdf==1.27.1
pypdfium2==5.5.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pyyaml==6.0.2
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
ruamel-yaml==0.19.1
safetensors==0.7.0
sentry-sdk==2.54.0
setuptools==82.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
tqdm==4.67.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
